scripts/raycast/requirements.txt
astroid==2.5.6
autopep8==1.5.5
certifi==2021.5.30
charset-normalizer==2.0.3
et-xmlfile==1.1.0
flake8==3.8.4
greenlet==1.1.0
idna==3.2
isort==5.8.0
jedi==0.17.2
lazy-object-proxy==1.6.0
mccabe==0.6.1
msgpack==1.0.2
neovim==0.3.1
numpy==1.20.3
openpyxl==3.0.7
parso==0.7.1
pluggy==0.13.1
protobuf==3.15.8
pycodestyle==2.6.0
pydocstyle==6.1.1
pyflakes==2.2.0
pylint==2.8.3
pynvim==0.4.3
python-dotenv==0.18.0
python-jsonrpc-server==0.4.0
python-language-server==0.36.2
requests==2.26.0
rope==0.19.0
six==1.15.0
snowballstemmer==2.1.0
toml==0.10.2
ujson==4.0.2
urllib3==1.26.6
wrapt==1.12.1
yapf==0.31.0
